refactor: pr 반영
