+++ src/app/_waiting/types/game.ts
