+++ src/components/app-bar/index.tsx
@@ -3,18 +3,13 @@ import { ArrowLeft } from "lucide-react"
 type Props = {
   title: string
   isBackButton?: boolean
-  callback?: () => void
+  onBackClick?: () => void
 }
 
-export default function AppBar({ title, isBackButton = false, callback }: Props) {
-  const onClick = () => {
-    if (typeof callback !== "undefined") {
-      callback()
-    }
-  }
+export default function AppBar({ title, isBackButton = false, onBackClick = () => {} }: Props) {
   return (
     <section className="relative p-[28px]">
-      {isBackButton && <ArrowLeft onClick={onClick} className="absolute" stroke="#98A2B3" />}
+      {isBackButton && <ArrowLeft onClick={onBackClick} className="absolute" stroke="#98A2B3" />}
       <span className="block text-center text-[14px] font-bold text-gray-400">{title}</span>
     </section>
   )
